refactor(back): cleanup user endpoint and fix security semantics

+++ back/src/main/java/com/openclassrooms/starterjwt/services/UserService.java
@@ -1,5 +1,6 @@
 package com.openclassrooms.starterjwt.services;
 
+import com.openclassrooms.starterjwt.exception.NotFoundException;
 import com.openclassrooms.starterjwt.models.User;
 import com.openclassrooms.starterjwt.repository.UserRepository;
 import org.springframework.stereotype.Service;
@@ -13,10 +14,13 @@ public class UserService {
     }
 
     public void delete(Long id) {
+        if (!this.userRepository.existsById(id)) {
+            throw new NotFoundException();
+        }
         this.userRepository.deleteById(id);
     }
 
     public User findById(Long id) {
-        return this.userRepository.findById(id).orElse(null);
+        return this.userRepository.findById(id).orElseThrow(NotFoundException::new);
     }
 }
